+++ app/faceanalysis/models/image_status_enum.py
