+++ app/wait-for-dependencies.sh
