+++ app/worker.py
@@ -1,6 +1,6 @@
-from api.models.database_manager import DatabaseManager
-from api.models.models import init_models
-from api.pipeline import Pipeline
+from faceanalysis.models.database_manager import DatabaseManager
+from faceanalysis.models.models import init_models
+from faceanalysis.pipeline import Pipeline
 
 if __name__ == "__main__":
     db = DatabaseManager()
