+++ app/wsgi.py
@@ -1,4 +1,4 @@
-from api.api import app as application
+from faceanalysis.api import app as application
 
 if __name__ == '__main__':
     application.run()
