+++ run-test.sh
@@ -12,8 +12,8 @@ trap cleanup EXIT
 docker-compose down
 docker-compose build --build-arg DEVTOOLS=true
 
-docker-compose run --rm --no-deps --entrypoint=python3 api -m pylint /app/api
-docker-compose run --rm --no-deps --entrypoint=python3 api -m flake8 /app/api
+docker-compose run --rm --no-deps --entrypoint=python3 api -m pylint /app/faceanalysis
+docker-compose run --rm --no-deps --entrypoint=python3 api -m flake8 /app/faceanalysis
 
 APP_PORT="${app_port}" \
 DATA_DIR="${data_dir}" \
